Rename convert_to_active.py to camelcase scheme
